Move `common.mjs` to `index/common.mjs`
