+++ packages/nhsuk-frontend/src/nhsuk/components/error-summary/error-summary.mjs
@@ -1,4 +1,4 @@
-import { getFragmentFromUrl } from '../../common.mjs'
+import { getFragmentFromUrl } from '../../common/index.mjs'
 import { Component } from '../../component.mjs'
 
 /**
